refactor: Service 빈은 각 테스트에서 등록하도록 수정
